% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_capacity_csvs.R
\name{generate_capacity_csvs}
\alias{generate_capacity_csvs}
\title{Generate Capacity CSVs using Plutus}
\usage{
generate_capacity_csvs(
  workdir,
  gcam_database_path,
  scenario_names = NULL,
  regions = NULL,
  output_dir = NULL
)
}
\arguments{
\item{workdir}{The working directory where outputs will be saved.}

\item{gcam_database_path}{The path to the GCAM database.}

\item{scenario_names}{A vector of scenario names to process. Default is \code{NULL}, meaning all scenarios in the database are selected.}

\item{regions}{A vector of region names to include in the analysis. Default is \code{NULL}, meaning all regions in the database are selected.}

\item{output_dir}{The directory where the output CSV files will be saved (default: 'generated_csvs' in \code{workdir}).}
}
\value{
The path to the directory where the generated CSV files are saved.
}
\description{
This function uses the \code{plutus} package to generate CSV files for capacity additions and retirements
based on the specified GCAM database and other input parameters. If the \code{plutus} package is not
installed, it will be installed automatically along with its dependencies.
}
