% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_intensities.R
\docType{data}
\name{job_intensities}
\alias{job_intensities}
\title{Job Intensities by Region and Technology}
\format{
A data frame with 3 columns:
\describe{
\item{region}{The GCAM region (e.g., "Africa_Eastern").}
\item{technology}{The power generation technology (e.g., "Solar").}
\item{jobs_per_GW}{Job intensity in Jobs/GW.}
}
}
\source{
Provided by the package author.
}
\usage{
job_intensities
}
\description{
A dataset containing job intensities in Jobs/GW for various regions and technologies.
}
\keyword{datasets}
