% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_visualization_app.R
\name{run_visualization_app}
\alias{run_visualization_app}
\title{Run Visualization App}
\usage{
run_visualization_app(results)
}
\arguments{
\item{results}{A named list containing data frames of the different result types.}
}
\description{
Launches a Shiny app for interactive exploration of the results.
}
