% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_results.R
\name{visualize_results}
\alias{visualize_results}
\title{Visualize Results}
\usage{
visualize_results(results, output_dir)
}
\arguments{
\item{results}{A named list containing data frames of the different result types.}

\item{output_dir}{A character string specifying the directory where the PNG files will be saved.}
}
\description{
Visualizes outputs of the \code{calculate_power_jobs} function and saves them as PNG files.
}
